+++ src/core/thread/network_diagnostic_tlvs.hpp
@@ -43,8 +43,7 @@
 #include "common/tlvs.hpp"
 #include "net/ip6_address.hpp"
 #include "radio/radio.hpp"
-#include "thread/device_mode.hpp"
-#include "thread/mle_constants.hpp"
+#include "thread/mle_types.hpp"
 
 namespace ot {
